晒单完成

+++ knock Honey/最新揭晓/宝贝详情/viewController/KHDetailViewController.m
@@ -163,7 +163,7 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath{
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
-    NSLog(@"%d",indexPath.row);
+    NSLog(@"%ld",(long)indexPath.row);
 }
 
 //夺宝中下面的，点击按钮
